+++ shell.nix
@@ -18,13 +18,18 @@ craneLib.devShell
         cargo-flamegraph
         rust-analyzer
 
+        pkg-config
+        openssl.dev
+
         # Our own scripts
         # FIXME: Broken
         # generate-coverage
-        playground
         update-wasm-app
         update-wasm-grammars
         verify-documented-usage
+      ]
+      ++ pkgs.lib.optionals (!stdenv.isDarwin) [
+        playground
       ];
     } else { })
   )