+++ topiary-cli/tests/sample-tester.rs
@@ -5,7 +5,7 @@ use topiary_core::test_utils::pretty_assert_eq;
 
 use tempfile::TempDir;
 
-fn io_test(file: &str) -> () {
+fn io_test(file: &str) {
     let input = PathBuf::from(format!("tests/samples/input/{file}"));
     let expected = PathBuf::from(format!("tests/samples/expected/{file}"));
 
@@ -68,7 +68,7 @@ fn input_output_tester() {
     io_test("toml.toml");
 
     #[cfg(feature = "tree_sitter_query")]
-    io_test("tree-sitter-query.scm");
+    io_test("tree_sitter_query.scm");
 }
 
 // Test that our query files are properly formatted