+++ topiary-config/src/lib.rs
@@ -119,7 +119,8 @@ impl Default for Configuration {
         let mut program = Program::<CacheImpl>::new_from_source(
             Source::Builtin
                 .read()
-                .expect("Evaluating the builtin configuration should be safe"),
+                .expect("Evaluating the builtin configuration should be safe")
+                .as_slice(),
             "builtin",
             std::io::empty(),
         )
@@ -166,4 +167,9 @@ impl From<SerdeConfiguration> for Configuration {
 
         Self { languages }
     }
+}
+
+pub(crate) fn project_dirs() -> directories::ProjectDirs {
+    directories::ProjectDirs::from("", "", "topiary")
+        .expect("Could not access the OS's Home directory")
 }